refactor(api): regroup /internal/* concern under shared/internal-routes/
The 5 files that gate /internal/* routes (signature primitives, signature middleware,
private-network middleware, env-driven composer, signed-fetch client) lived split across
shared/ root and shared/middleware/. Group them under shared/internal-routes/ so the concern
reads as one folder.

shared/middleware/ now holds only cross-cutting Hono middlewares (auth, error, logger, org).
CLAUDE.md (api + shared), ROADMAP and docs/CRON updated; knip entry rewired.
