+++ ROADMAP.md
@@ -877,7 +877,7 @@ The **Billing** section above lays the foundation: `PLANS` config, `useEntitleme
 - [x] **Front step 2 — collapse `adapters/` + `common/` + `providers/` into `shared/`**. `shared/api/`, `shared/auth/`, `shared/components/`, `shared/app-providers.tsx`, `shared/env.ts`, `shared/utils.ts`. 4 sub-folders + 2 root files; lean.
 - [x] **Front step 3 — code-based routing** (Option C). `routes/` folder + `routeTree.gen.ts` + `@tanstack/router-plugin` deleted. Each feature exposes `<name>Route(parent)` factory in `<name>.route.tsx`; `apps/app/src/router.tsx` defines layouts/gates inline + assembles via `addChildren`. Library features (`security/`, `rgpd/`) stay route-less, composed by `account/`. Feature-scoped queries/mutations relocation to `features/<x>/api/` deferred (current `shared/api/queries+mutations/` remains pragmatic until duplication justifies the split).
 - [x] **Back step 1 — pivot `apps/api/src/` to `modules/<context>/`**: `modules/uploads/`, `modules/rgpd/` shipped (auth + organizations stay at `apps/api/src/auth.ts` — BetterAuth singleton, no DDD). Each contains `application/{dto,services,ports for module-private interfaces}` + `infrastructure/{services,repositories}` + `routes.ts` (where applicable) + `module.ts`. `adapters/` removed; `infrastructure/` is the DDD-canonical naming. **`modules/email/` deliberately not created** — email is pure infra (no domain, no use cases, no routes), lives in `shared/services/email.service.ts` consuming `shared/ports/email.port.ts` (rule "shared kernel" in CLAUDE.md).
-- [x] **Back step 2 — extract `shared/`**: `apps/api/src/shared/{middleware,ports,services}/` + `{env,logger,internal-signature,internal-fetch}.ts`. Cross-context port interfaces (`IStorageService`, `IEmailService`) live in `shared/ports/`; cross-context impls (`ResendEmailService`) in `shared/services/`. The composed env-driven gate for `/internal/*` lives in `shared/middleware/internal-layers.ts` (single source — any future module exposing `/internal/*` consumes the same `internalLayers`). `common/` deleted.
+- [x] **Back step 2 — extract `shared/`**: `apps/api/src/shared/{middleware,ports,services}/` + process-level singletons at the root. Cross-context port interfaces (`IStorageService`, `IEmailService`) live in `shared/ports/`; cross-context impls (`ResendEmailService`) in `shared/services/`. The composed env-driven gate for `/internal/*` lives in `shared/internal-routes/internal-layers.ts` (single source — any future module exposing `/internal/*` consumes the same `internalLayers`); the whole `/internal/*` concern (signature primitives + middlewares + signed-fetch client) is grouped under `shared/internal-routes/`. `common/` deleted.
 - [x] **Back step 3 — `module.ts` per context (split form)**: each `module.ts` defines an inwire `defineModule()` (typed prerequisites local to the module). `di/container.ts` chains `.addModule(emailModule).addModule(uploadsModule).addModule(rgpdModule)`. Routes stay in `routes.ts` / `internal.routes.ts` and are imported by `index.ts` directly — splitting DI from routes avoids a `module.ts → routes.ts → di/container.ts → module.ts` cycle (Biome `noImportCycles` flagged it; the SOTA is to keep DI wiring and route mounting on separate import graphs).
 - [ ] **Back step 4 — split DB schema**: today `packages/drizzle/src/schema/auth.ts` is the single file (auth + multi-tenant + rgpd fields all bundled). Split into `packages/drizzle/src/schema/<context>.ts` (auth, organization, uploads when keys land in DB, rgpd, …), barrel `index.ts` re-exports. Each module owns its tables. Removing a module = remove the `export *` line + revert the migration. Last front of the vertical-slice migration; ship before Phase B (Billing) so the new `subscription` / `payment` tables land in their own file from day one.
 - [ ] **Removability dry-run** on the smallest module (probably `rgpd` since fresh in memory, or `uploads` if smaller) — delete it end-to-end, run `pnpm ci:check`, document the diff in `docs/HISTORY.md` as the canonical "how to remove a feature" example.
