+++ apps/api/src/shared/CLAUDE.md
@@ -4,12 +4,16 @@ Loaded when working inside `apps/api/src/shared/`. Cross-cutting infra placement
 
 ## What lives here
 
-- `middleware/` — auth, error, logger, internal-signature, private-network, org, internal-layers (env-gate composing internal middlewares for `/internal/*`)
+- `middleware/` — cross-cutting Hono middlewares: `auth`, `error`, `logger`, `org`. Internal-route gating lives in its own folder (see below).
+- `internal-routes/` — everything that gates `/internal/*` (cron callers, GH Actions, sidecar schedulers). Grouped by concern, not by technical type:
+  - `internal-signature.ts` — HMAC primitives (canonicalize/sign/verify)
+  - `internal-signature.middleware.ts` — server-side verifier (`requireInternalSignature`)
+  - `private-network.middleware.ts` — RFC1918/loopback gate (`requirePrivateNetwork`)
+  - `internal-layers.ts` — env-driven composer (`INTERNAL_AUTH_LAYERS`) — modules use `internalLayers` as one spread
+  - `internal-fetch.ts` — client-side `signedInternalFetch`, importable by external schedulers calling `/internal/*`
 - `ports/` — cross-context port interfaces (consumed by 2+ contexts, OR pure transport). Currently: `email.port`, `storage.port`, `outbox.port`, `audit.port`.
 - `services/` — cross-context port impls. Currently: `ResendEmailService`, `DrizzleOutboxRepository`, `DrizzleAuditRepository`, `OutboxDispatcher` (LISTEN/NOTIFY worker), `AuditEventSubscriber` + `WebhookFanoutSubscriber` (built-in outbox subscribers).
 - `env.ts`, `logger.ts` — process-level singletons
-- `internal-signature.ts` — HMAC primitives (canonicalize/sign/verify)
-- `internal-fetch.ts` — `signedInternalFetch`, importable by external schedulers (cron workers, GH Actions) calling `/internal/*`
 - `transaction.ts` — `type ITransaction = Transaction` (Drizzle alias). Type-only swap-point exception to "no infra in app layer" rule.
 - `event-emitter.ts` — `emitEvent(outbox, ...)` helper for code that emits events outside an aggregate flow (BetterAuth bridge, RGPD service, uploads). Use this instead of `outbox.enqueue` directly to keep the source/scope shape consistent.
 - `aead.ts` — XChaCha20-Poly1305 encrypt/decrypt + HKDF per-org sub-key for webhook secrets at rest.
