+++ apps/api/src/shared/__TESTS__/internal-signature.middleware.test.ts
@@ -1,8 +1,13 @@
 import { describe, expect, it } from "bun:test";
 import { Hono } from "hono";
 import { env } from "../env";
-import { buildSignatureHeader, canonicalize, SIGNATURE_HEADER, sign } from "../internal-signature";
-import { requireInternalSignature } from "../middleware/internal-signature.middleware";
+import {
+  buildSignatureHeader,
+  canonicalize,
+  SIGNATURE_HEADER,
+  sign,
+} from "../internal-routes/internal-signature";
+import { requireInternalSignature } from "../internal-routes/internal-signature.middleware";
 
 const KEY = env.INTERNAL_SIGNING_KEY ?? "";
